requirements.txt
babel==2.16.0
itsdangerous==2.2.0
bcrypt==4.2.0
cryptography==44.0.0
email-validator==2.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.5
openai==1.54.3
opencage==3.0.3
packaging==24.2
pip==25.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
WTForms==3.2.1
psycopg2==2.9.10   #for PostgreSQL connection